requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
sqlmodel==0.0.22
alembic==1.14.0
celery==5.4.0
redis==5.2.1
scikit-learn==1.6.0
scipy==1.14.1
numpy==2.2.0
pandas==2.2.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
tenacity==9.0.0
